Test.QuickCheck.Arbitrary.ADT

Description

Type classes to assist random generation of values for various types of abstract data types.

How to use this library

How to use ToADTArbitrary with Generic.

{-# LANGUAGE DeriveGeneric #-}

import Data.Proxy
import GHC.Generics
import Test.QuickCheck
import Test.QuickCheck.Arbitrary.ADT

-- Sum Type, multiple constructors with parameters
data Fruit
  = Apple Int
  | Orange String Int
  | PassionFruit Int String Int
  deriving (Generic, Show)

-- Product Type, single constructor
data Person =
  Person
    { name :: String
    , age  :: Int
    } deriving (Generic, Show)

Any type that implements ToADTArbitrary must also implement Arbitrary. These examples all require that the data type is an instance of Generic.

instance Arbitrary Fruit where
  arbitrary = genericArbitrary

instance ToADTArbitrary Fruit

instance Arbitrary Person where
  arbitrary = genericArbitrary

instance ToADTArbitrary Person

Now we can use toADTArbitrarySingleton to produce an arbitrary value of one random constructor along with some metadata. toADTArbitrary will produce an arbitrary value for each constructor and return it along with a String of the constructor name.

λ> generate (toADTArbitrarySingleton (Proxy :: Proxy Fruit))
ADTArbitrarySingleton {
  adtasModuleName = "Ghci1"
, adtasTypeName = "Fruit"
, adtasCAP = ConstructorArbitraryPair {
    capConstructor = "Apple", capArbitrary = Apple 30}}

λ> generate (toADTArbitrary (Proxy :: Proxy Fruit))
ADTArbitrary {
  adtModuleName = "Ghci1"
, adtTypeName = "Fruit"
, adtCAPs = [
    ConstructorArbitraryPair {
        capConstructor = "Apple"
      , capArbitrary = Apple 17}
  , ConstructorArbitraryPair {
        capConstructor = "Orange"
      , capArbitrary = Orange "abcdef" 18}
  , ConstructorArbitraryPair {
        capConstructor = "PassionFruit"
      , capArbitrary = PassionFruit 16 "datadata" 6}]}

λ> generate (toADTArbitrarySingleton (Proxy :: Proxy Person))
ADTArbitrarySingleton {
  adtasModuleName = "Ghci1"
, adtasTypeName = "Person"
, adtasCAP = ConstructorArbitraryPair {capConstructor = "Person", capArbitrary = Person {name = "John Doe", age = 30}}}

λ> generate (toADTArbitrary (Proxy :: Proxy Person))
ADTArbitrary {
  adtModuleName = "Ghci1"
, adtTypeName = "Person"
, adtCAPs = [ConstructorArbitraryPair {capConstructor = "Person", capArbitrary = Person {name = "Jane Doe", age = 15}}]}

Data types

data ConstructorArbitraryPair a

ConstructorArbitraryPair holds the construct name as a string and an arbitrary instance of that constructor.

Constructors

ConstructorArbitraryPair

Fields

capConstructor :: String

capArbitrary :: a

data ADTArbitrarySingleton a

ADTArbitrarySingleton holds the type name and one ConstructorArbitraryPair.

Constructors

ADTArbitrarySingleton

Fields

adtasModuleName :: String

adtasTypeName :: String

adtasCAP :: ConstructorArbitraryPair a

data ADTArbitrary a

ADTArbitrary holds the type name and a ConstructorArbitraryPair for each constructor.

Constructors

ADTArbitrary

Fields

adtModuleName :: String

adtTypeName :: String

adtCAPs :: [ConstructorArbitraryPair a]

Type classes

class ToADTArbitrary a where

ToADTArbitrary generalizes the production of arbitrary values for Sum types. and Product types.

Minimal complete definition

Nothing

Methods

toADTArbitrarySingleton :: Proxy a -> Gen (ADTArbitrarySingleton a)

produce an arbitrary instance of one random constructor

toADTArbitrarySingleton :: (Generic a, GToADTArbitrarySingleton (Rep a)) => Proxy a -> Gen (ADTArbitrarySingleton a)

produce an arbitrary instance of one random constructor

toADTArbitrary :: Proxy a -> Gen (ADTArbitrary a)

produce an arbitrary instance for each constructor in type a.

toADTArbitrary :: (Generic a, GToADTArbitrary (Rep a)) => Proxy a -> Gen (ADTArbitrary a)

produce an arbitrary instance for each constructor in type a.

Generic type classes

class GToADTArbitrarySingleton rep where

GToADTArbitrarySingleton creates an arbitrary value and returns the name of the constructor that was used to create it and the type name.

Methods

gToADTArbitrarySingleton :: Proxy rep -> Gen (ADTArbitrarySingleton (rep a))

class GToADTArbitrary rep where

GToADTArbitrary is a typeclass for generalizing the creation of a list of arbitrary values for each constructor of a type. It also returns the name of the constructor and the type name for reference and file creation.

Methods

gToADTArbitrary :: Proxy rep -> Gen (ADTArbitrary (rep a))

class GArbitrary rep where

GArbitrary is a typeclass for generalizing the creation of single arbitrary product and sum types. It creates an arbitrary generating function of this style: TypeName <$> arbitrary <*> arbitrary.

Methods

gArbitrary :: Gen (rep a)

genericArbitrary :: (Generic a, GArbitrary (Rep a)) => Gen a

Create a arbitrary generator for a specified a type in a naive way. Please be careful when using this function, particularly for recursive types.